Make runReport generic: emit entries, consumers map to rows
runReport now hands each raw TsMigratingReportEntry to `onEntry` and lets the
consumer decide how to represent it, instead of producing a ReportRow itself.
`toReportRow` (+ ReportRow + tests) moves to its own helper that json/ndjson
call. The engine no longer knows about the JSON row shape — it only iterates,
tallies the gate, and exits — so it can back any reporter (the pretty reporter
could adopt it too). runReport also returns the tally for building summaries.
